requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
redis==5.0.0
celery==5.3.4
requests==2.31.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.5.0
whisper==1.0
google-generativeai==0.2.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pyttsx3==2.90
googletrans==4.0.0rc1
pytest==7.4.3
pytest-flask==1.3.0
bcrypt==4.1.2
itsdangerous==2.1.2
gunicorn==21.2.0
pandas==2.1.4
psutil==5.9.7